100% parsing OK on my testcases

+++ tests/run_test.rb
@@ -2,24 +2,27 @@ require_relative "../lib/crokus"
 
 puts "running parser tests"
 
-pp_c_files=Dir["*_pp.c"]
-puts "suppressing generated _pp.c files : #{pp_c_files.size}"
-pp_c_files.each do |c|
-  system("rm -f #{c}")
+unless ARGV.first=="--all"
+  pp_c_files=Dir["*_pp.c"]
+  puts "suppressing generated _pp.c files : #{pp_c_files.size}"
+  pp_c_files.each do |c|
+    system("rm -f #{c}")
+  end
 end
 
 compiler=Crokus::Compiler.new
 compiler.options[:mute]=true
-files=Dir["*.c"]
+files=Dir["*.c"].sort
 nb_tests=files.size
-max_length=files.max_by{|s| s.length}.length+3
+max_length=files.max_by{|s| s.length}.length+4
 success=0
+
 files.each do |c|
   begin
-    print "parsing "+(" "+c).rjust(max_length,".")
+    print "parsing "+(c+" ").ljust(max_length,".")
     result=compiler.compile c
     status=result==true ? "OK": "nok"
-    puts status.rjust(10,' ')
+    puts status.rjust(3)
     success+=1
   rescue Exception
     puts
